+++ website/src/pages/Documentation/hooks/useStateTogetherWithPerUserValues.tsx
@@ -135,7 +135,7 @@ const api = (
       id='options'
       items={[
         {
-          name: 'omitLocalValue',
+          name: 'omitMyValue',
           type: 'boolean',
           default: 'false',
           description: <Markdown>If this option is `true`, the local value will not be included in the `allValues` object.</Markdown>,