+++ packages/react-together/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Unpublished
- - Added `omitLocalValue` option flag in `useStateTogetherWithPerUserValues`
+ - Added `omitMyValue` option flag in `useStateTogetherWithPerUserValues`
 
 ## 0.3.0
 `2024-12-13`